+++ src/Query/Where.php
@@ -112,7 +112,7 @@ final class Where
             return \sprintf('%s IS NULL', $column);
         }
 
-        $comparator = match ($specification->sign()) {
+        return match ($specification->sign()) {
             Sign::in => $this->buildInSql($driver, $specification),
             default => \sprintf(
                 '%s %s ?',
@@ -120,19 +120,6 @@ final class Where
                 $sign,
             ),
         };
-
-        if (
-            $driver === Driver::sqlite &&
-            \in_array(
-                $specification->sign(),
-                [Sign::startsWith, Sign::endsWith, Sign::contains],
-                true,
-            )
-        ) {
-            $comparator .= " ESCAPE '\\'";
-        }
-
-        return $comparator;
     }
 
     private function buildComposite(